Trait r3::bind::Binder

pub trait Binder {
    type Runtime: RuntimeBinder;

    fn register_dependency(&self, ctx: &mut CfgBindCtx<'_>);
    fn into_runtime_binder(self) -> Self::Runtime;
}

Represents a binder, which represents a specific way to access the contents of a binding from a runtime function.

See the module-level documentation for more.

Stability

This trait is covered by the application-side API stability guarantee with a few exceptions, which are documented on a per-item basis.

Required Associated Types

type Runtime: RuntimeBinder

The runtime representation of Self.

Stability

This method is unstable.

Required Methods

fn register_dependency(&self, ctx: &mut CfgBindCtx<'_>)

Define a binding dependency in CfgBindCtx::bind_registry.

Stability

This method is unstable.

fn into_runtime_binder(self) -> Self::Runtime

Convert self to the runtime representation (RuntimeBinder).

Stability

This method is unstable.