Trait r3::bind::ExecutableDefiner

pub unsafe trait ExecutableDefiner: Sized + Sealed {
    // Required method
    fn start(self, start: Closure) -> Self;
}

A trait for definer objects (static builders) for kernel objects that can spawn a thread that executes after the execution of all startup hooks is complete.

See ExecutableDefinerExt for an extension trait providing a method to specify a closure that consumes binders.

Safety

At any point of time, the provided Closure must never be invoked by two threads simultaneously. It can be called for multiple times, however.

Required Methods

fn start(self, start: Closure) -> Self

Use the specified function as the entry point of the executable object being defined.