Trait r3::bind::FnBind

pub trait FnBind<Binder> {
    type Output: 'static;
    type BoundFn: 'static + FnOnce() + Copy + Send
    where
        <Self::BoundFn as FnOnce()>::Output == Self::Output;

    fn bind(self, binder: Binder, ctx: &mut CfgBindCtx<'_>) -> Self::BoundFn;
}

A trait for closures that can receive bindings materialized through specific binders (Binder).

FnBind<(B0, B1, ...)> is implemented for impl for<'call> FnOnce(M0<'call>, M1<'call>, ...) + Copy + Send + 'static, where Mn<'call> == Bn::Runtime::Target<'call> (Bn’s materialized form).

Stability

This trait is covered by the application-side API stability guarantee with the exception of its members, which are implementation details.

Required Associated Types

type Output: 'static

type BoundFn: 'static + FnOnce() + Copy + Send
where
<Self::BoundFn as FnOnce()>::Output == Self::Output

Required Methods

fn bind(self, binder: Binder, ctx: &mut CfgBindCtx<'_>) -> Self::BoundFn