Module r3::hunk

Type-safe hunks

Structs

CfgHunkBuilder
Configuration builder type for Hunk.

DefaultInitTag
As a generic parameter of CfgHunkBuilder, indicates that the hunk should be initialized with Init.

Hunk
Represents a single typed hunk in a system.

ZeroInitTag
As a generic parameter of CfgHunkBuilder, indicates that the hunk should be zero-initialized.

Constants

INIT_HOOK_PRIORITY
The priority of the startup hooks used to initialize typed hunks. It has a negative value so that startup hooks with non-negative priorities (which can be created without unsafe blocks) will never see an uninitialized value in a typed hunk.

Traits

HunkIniter
Implemented on DefaultInitTag and ZeroInitTag when T can be initialized in this way.