Struct r3::hunk::Hunk

pub struct Hunk<System, T: ?Sized> { /* fields omitted */ }

Represents a single typed hunk in a system.

Hunks are nothing more than static variables defined in a kernel configuration. They come in handy when you are designing a component that can be instantiated by a kernel configuration and wanting each instance to have its own separate state data.

This type is implemented on top of r3::kernel::Hunk, the untyped hunk type.

Implementations

impl<System: Kernel, T: ?Sized> Hunk<System, T>

pub const fn build() -> CfgHunkBuilder<System, T, DefaultInitTag>

Construct a CfgTaskBuilder to define a hunk in a configuration function.

impl<System, T: ?Sized> Hunk<System, T>

pub const unsafe fn transmute<U>(self) -> Hunk<System, U>

Reinterpret the hunk as another type.

Safety

Similarly to core::mem::transmute, this is incredibly unsafe.

The byte offset must be valid for the destination type.

impl<System: Kernel, T: ?Sized> Hunk<System, T>

pub fn untyped_hunk(this: Self) -> Hunk<System>

Get the untyped hunk.

pub fn as_ptr(this: Self) -> *const T

Get a raw pointer to the hunk’s contents.

pub fn as_bytes_ptr(this: Self) -> *const [u8]

Get a raw pointer to the raw bytes of the hunk.

pub unsafe fn as_bytes(this: Self) -> &'static [u8]

Get a reference to the raw bytes of the hunk.

Safety

The result might include uninitialized bytes and/or interior mutability, so it might be unsafe to access.