Struct r3::kernel::cfg::Cfg

pub struct Cfg<'c, C> where
    C: CfgBase, { /* private fields */ }

Wraps a raw_cfg::CfgBase to provide higher-level services.

Implementations

impl<'c, C> Cfg<'c, C> where
C: CfgBase,

pub const fn raw(&mut self) -> &mut C

Mutably borrow the underlying C.

pub const fn num_task_priority_levels(&mut self, new_value: usize) where
C: CfgBase,

Specify the number of task priority levels.

The RAM consumption by task ready queues is expected to be proportional to the number of task priority levels. In addition, the scheduler may be heavily optimized for the cases where the number is very small (e.g., < 16). This optimization can provide a significant performance improvement if the target processor does not have a CTZ (count trailing zero) instruction, barrel shifter, or hardware multiplier.

Kernels may set an arbitrary upper bound for the number of task priority levels.

pub const fn finish_phase1(&mut self) -> CfgPhase1Data<<C as CfgBase>::System> where
C: CfgBase,

Finalize self for the phase 1 configuration.

This method constructs CfgPhase1Data, which must be passed to attach_phase1! to proceed with the kernel-independent configuration process.

pub const fn finish_phase2(&mut self) -> CfgPhase2Data<<C as CfgBase>::System> where
C: CfgBase,

Finalize self for the phase 2 configuration.

This method constructs CfgPhase2Data, which must be passed to attach_phase2! to proceed with the kernel-independent configuration process.

pub const fn finish_phase3_interrupt(&mut self) where
C: CfgInterruptLine,
<C as CfgBase>::System: CfgPhase2<<C as CfgBase>::System>,
<C as CfgBase>::System: KernelInterruptLine,

Perform additional finalization tasks for interrupt line configuration.

This method must be called before Self::finish_phase3 if C implements CfgInterruptLine.

pub const fn finish_phase3(self) -> CfgPhase3Data<<C as CfgBase>::System> where
C: CfgBase,
<C as CfgBase>::System: CfgPhase2<<C as CfgBase>::System>,

Finalize self for the phase 3 configuration.

This method makes the last set of changes to the referenced C: impl CfgBase. It also constructs CfgPhase3Data, which must be passed to attach_phase3! to complete the kernel-independent configuration process.