Enum r3::kernel::LockMutexError

pub enum LockMutexError {
    NoAccess,
    BadContext,
    Interrupted,
    WouldDeadlock,
    BadParam,
    Abandoned,
}

Error type for Mutex::lock.

See ResultCode for all result codes and generic descriptions.

Variants

NoAccess

Invalid object access.

BadContext

CPU Lock is active, or the current context is not waitable.

Interrupted

WouldDeadlock

The current task already owns the mutex.

BadParam

The mutex was created with the protocol attribute having the value Ceiling and the current task’s priority is higher than the mutex’s priority ceiling.

Abandoned

The previous owning task exited while holding the mutex lock. The current task shall hold the mutex lock, but is up to make the state consistent.