Enum r3::kernel::MutexProtocol

pub enum MutexProtocol {
    None,
    Ceiling(usize),
}

Specifies the locking protocol to be followed by a mutex.

Relation to Other Specifications: The operating systems and operating system specifications providing an interface for specifying a mutex protocol include (but are not limited to) the following: POSIX (pthread_mutexattr_setprotocol and PTHREAD_PRIO_PROTECT, etc.), RTEMS Classic API (RTEMS_PRIORITY_CEILING, etc.), and μITRON4.0 (TA_CEILING, etc.).

Rationale: When this enumerate type was added, the plan was to only support the priority ceiling protocol, so having a method CfgMutexBuilder::ceiling_priority taking a priority ceiling value would have been simpler. Nevertheless, it was decided to use this enumerate type to accomodate other protocols in the future and to allow specifying protocol-specific parameters.

Variants

None

Locking the mutex does not affect the priority of the owning task.

Ceiling(usize)

Locking the mutex raises the effective priority of the owning task to the mutex’s priority ceiling according to the immediate priority ceiling protocol. The inner value specifies the priority ceiling.

The value must be in range 0..num_task_priority_levels.