Enum r3::kernel::ResultCode

#[repr(i8)]pub enum ResultCode {
    Success,
    NotSupported,
    BadParam,
    BadId,
    BadContext,
    NotOwner,
    WouldDeadlock,
    BadObjectState,
    QueueOverflow,
    Abandoned,
    Interrupted,
    Timeout,
}

All result codes (including success) that the C API can return.

Relation to Other Specifications: All error codes are intentionally matched to their equivalents in μITRON4.0 for no particular reasons.

Rationale: Using the C API result codes internally reduces the interop overhead at an API surface.

Variants

Success

The operation was successful. No additional information is available.

NotSupported

The operation is not supported.

BadParam

A parameter is invalid in a way that is no covered by any other error codes.

BadId

A specified object identifier (Id) is invalid.

BadContext

The current context disallows the operation.

NotOwner

The caller does not own the resource.

WouldDeadlock

Resource deadlock would occur.

BadObjectState

A target object is in a state that disallows the operation.

QueueOverflow

An operation or an object couldn’t be enqueued because there are too many of such things that already have been enqueued.

Abandoned

The owner of a mutex exited while holding the mutex lock.

Interrupted

The wait operation was interrupted by Task::interrupt.

Timeout

The operation timed out.

Implementations

impl ResultCode

pub fn as_str(self) -> &'static str

Get the short name of the result code.

Examples

use r3::kernel::ResultCode;
assert_eq!(ResultCode::BadObjectState.as_str(), "BadObjectState");

impl ResultCode

pub fn is_err(self) -> bool

Get a flag indicating whether the code represents a failure.

Failure codes have negative values.

pub fn is_ok(self) -> bool

Get a flag indicating whether the code represents a success.

Success codes have non-negative values.