Struct r3::kernel::event_group::EventGroup

pub struct EventGroup<System>(_)
where
         System: NotSupportedYet;

Represents a single owned event group in a system.

This type is ABI-compatible with System::RawEventGroupId.

See EventGroupRef for the borrowed counterpart. See EventGroupMethods for the operations provided by this handle type.

An event group is a set of bits that can be updated and waited for to be set.

Relation to Other Specifications: event group (FreeRTOS), event group (Freescale MQX), EventFlags (Mbed OS), events (OSEK/VDX, assigned to each extended task), event (RT-Thread), event set (RTEMS, assigned to each task), Eventflag (μITRON4.0)