Trait r3::kernel::raw_cfg::CfgBase

pub unsafe trait CfgBase {
    type System: KernelBase;

    fn num_task_priority_levels(&mut self, new_value: usize);
    fn startup_hook_define(&mut self, func: fn());
}

The trait for all kernel-specific low-level configurator types, used by the kernel static configuration process.

Safety

See the module documentation.

Stability

See the module documentation.

Required Associated Types

type System: KernelBase

Required Methods

fn num_task_priority_levels(&mut self, new_value: usize)

fn startup_hook_define(&mut self, func: fn())

Register a combined startup hook.

The configuration system calls this exactly once for each built system.