Struct r3::kernel::Hunk

pub struct Hunk<System> { /* fields omitted */ }

Represents a single hunk in a system.

Hunks are nothing more than static variables defined in a kernel configuration. They come in handy when you are designing a component that can be instantiated by a kernel configuration and wanting each instance to have its own separate state data.

This Hunk is untyped and only contains a starting address. See r3::hunk::Hunk for a type-safe wrapper of Hunk.

Relation to Other Specifications: None. The need for programmatically defining static regions has been traditionally fulfilled by code generation and preprocessor-based composition. The closest thing might be cell internal variables in the component system that comes with the TOPPERS kernels.

Implementations

impl<System: Port> Hunk<System>

pub const fn build() -> CfgHunkBuilder<System>

Construct a CfgTaskBuilder to define a hunk in a configuration function.

impl<System> Hunk<System>

pub const fn offset(self) -> usize

Get the offset of the hunk.

impl<System: Kernel> Hunk<System>

pub fn as_ptr(self) -> *mut u8

Get a raw pointer to the hunk’s contents.