Struct r3::kernel::Semaphore

#[repr(transparent)]pub struct Semaphore<System>(_, _);

Represents a single semaphore in a system.

A semaphore maintains a set of permits that can be acquired (possibly blocking) or released by application code. The number of permits held by a semaphore is called the semaphore’s value and represented by SemaphoreValue.

This type is ABI-compatible with Id.

Relation to Other Specifications: Present in almost every real-time operating system.

Implementations

impl<System: Port> Semaphore<System>

pub const fn build() -> CfgSemaphoreBuilder<System>

Construct a CfgTaskBuilder to define a semaphore in a configuration function.

impl<System> Semaphore<System>

pub const unsafe fn from_id(id: Id) -> Self

Construct a Semaphore from Id.

Safety

The kernel can handle invalid IDs without a problem. However, the constructed Semaphore may point to an object that is not intended to be manipulated except by its creator. This is usually prevented by making Semaphore an opaque handle, but this safeguard can be circumvented by this method.

pub const fn id(self) -> Id

Get the raw Id value representing this semaphore.

impl<System: Kernel> Semaphore<System>

pub fn drain(self) -> Result<(), DrainSemaphoreError>

Remove all permits held by the semaphore.

pub fn get(self) -> Result<SemaphoreValue, GetSemaphoreError>

Get the number of permits currently held by the semaphore.

pub fn signal(self, count: SemaphoreValue) -> Result<(), SignalSemaphoreError>

Release count permits, returning them to the semaphore.

pub fn signal_one(self) -> Result<(), SignalSemaphoreError>

Release a permit, returning it to the semaphore.

pub fn wait_one(self) -> Result<(), WaitSemaphoreError>

Acquire a permit, potentially blocking the calling thread until one is available.

This system service may block. Therefore, calling this method is not allowed in a non-waitable context and will return Err(BadContext).

Rationale: Multi-wait (waiting for more than one permit) is not supported because it introduces additional complexity to the wait queue mechanism by requiring it to reevaluate wait conditions after reordering the queue.

The support for multi-wait is relatively rare among operating systems. It’s not supported by POSIX, RTEMS, TOPPERS, VxWorks, nor Win32. The rare exception is μT-Kernel.

pub fn wait_one_timeout(
self,
timeout: Duration
) -> Result<(), WaitSemaphoreTimeoutError>

wait_one with timeout.

pub fn poll_one(self) -> Result<(), PollSemaphoreError>

Non-blocking version of wait_one. Returns immediately with PollSemaphoreError::Timeout if the unblocking condition is not satisfied.