Struct r3::kernel::StackHunk

pub struct StackHunk<System>(_, _);

Hunk for a task stack.

Implementations

impl<System> StackHunk<System>

pub const unsafe fn from_hunk(hunk: Hunk<System>, len: usize) -> Self

Construct a StackHunk from Hunk.

Safety

The caller is responsible for making sure the region represented by hunk is solely used for a single task’s stack.

Also, hunk must be properly aligned for a stack region.

pub fn into_inner(self) -> (Hunk<System>, usize)

Get the inner Hunk and the length, consuming self.

impl<System: Kernel> StackHunk<System>

pub fn as_ptr(&self) -> *mut [u8]

Get a raw pointer to the hunk’s contents.