Trait r3::kernel::KernelCfg1

pub unsafe trait KernelCfg1: Sized + Send + Sync + 'static {
    type TaskPriority: BinUInteger;

    const NUM_TASK_PRIORITY_LEVELS: usize;
    const TASK_PRIORITY_LEVELS: &'static [Self::TaskPriority];
}

Associates “system” types with kernel-private data. Use build! to implement.

Customizable things needed by both of Port and KernelCfg2 should live here because Port cannot refer to an associated item defined by KernelCfg2.

Safety

This is only intended to be implemented by build!.

Associated Types

type TaskPriority: BinUInteger

Unsigned integer type capable of representing the range 0..NUM_TASK_PRIORITY_LEVELS.

Loading content...
Associated Constants

const NUM_TASK_PRIORITY_LEVELS: usize

The number of task priority levels.

const TASK_PRIORITY_LEVELS: &'static [Self::TaskPriority]

All possible values of TaskPriority.

TASK_PRIORITY_LEVELS[i] is equivalent to TaskPriority::try_from(i).unwrap() except that the latter doesn’t work in const fn.

Loading content...