Module r3::kernel::traits

Re-exports all traits defined under this module for convenience.

Traits

CfgBase

The trait for all kernel-specific low-level configurator types, used by the kernel static configuration process.

CfgEventGroup

A low-level configurator trait providing a method to define an event group in the kernel static configuration process.

CfgInterruptLine

A low-level configurator trait providing a method to define an interrupt line in the kernel static configuration process.

CfgMutex

A low-level configurator trait providing a method to define a mutex in the kernel static configuration process.

CfgSemaphore

A low-level configurator trait providing a method to define a semaphore in the kernel static configuration process.

CfgStatic

A subtrait of raw_cfg::CfgBase that implies Self::System: KernelStatic.

CfgTask

A low-level configurator trait providing a method to define a task in the kernel static configuration process.

CfgTimer

A low-level configurator trait providing a method to define a timwer in the kernel static configuration process.

EventGroupHandle

The trait for safe wrappers of System::RawEventGroupId, i.e., EventGroup and EventGroupRef.

EventGroupMethods

The supported operations on EventGroupHandle.

Kernel

Provides access to the global functionalities of a kernel.

KernelAdjustTime

Provides the adjust_time method.

KernelBase

Provides access to the minimal API exposed by a kernel.

KernelBoostPriority

Provides the boost_priority method.

KernelEventGroup

Provides access to the event group API exposed by a kernel.

KernelInterruptLine

Provides access to the interrupt line API exposed by a kernel.

KernelMutex

Provides access to the mutex API exposed by a kernel.

KernelSemaphore

Provides access to the semaphore API exposed by a kernel.

KernelStatic

Associates static data to a system type.

KernelTaskSetPriority

Provides the task_set_priority method.

KernelTime

Provides the time method.

KernelTimer

Provides access to the timer API exposed by a kernel.

MutexHandle

The trait for safe wrappers of System::RawMutexId, i.e., Mutex and MutexRef.

MutexMethods

The supported operations on MutexHandle.

SemaphoreHandle

The trait for safe wrappers of System::RawSemaphoreId, i.e., Semaphore and SemaphoreRef.

SemaphoreMethods

The supported operations on SemaphoreHandle.

TaskHandle

The trait for safe wrappers of System::RawTaskId, i.e., Task and TaskRef.

TaskMethods

The supported operations on TaskHandle.

TimerHandle

The trait for safe wrappers of System::RawTimerId, i.e., Timer and TimerRef.

TimerMethods

The supported operations on TimerHandle.