Module r3::sync::mutex

Structs

Builder
Configuration builder type for Mutex.

Mutex
A mutual exclusion primitive useful for protecting shared data from concurrent access.

MutexGuard
An RAII implementation of a “scoped lock” of a mutex. When this structure is dropped, the lock will be released.

Enums

LockError
Error type of Mutex::lock.

MarkConsistentError
Error type of Mutex::mark_consistent.

TryLockError
Error type of Mutex::try_lock.

Type Definitions

LockResult
Type alias for the result of Mutex::lock.

TryLockResult
Type alias for the result of Mutex::try_lock.