Struct r3::sync::mutex::GenericMutex

pub struct GenericMutex<Cell, Mutex> { /* private fields */ }

Available on crate feature sync only.

A mutual exclusion primitive useful for protecting shared data from concurrent access.

This type is implemented using r3::kernel::Mutex, the low-level synchronization primitive and therefore inherits its properties. The important inherited properties are listed below:

When trying to lock an abandoned mutex, the lock function will return Err(LockError::Abandoned(lock_guard)). This state can be exited by calling GenericMutex::mark_consistent.

Mutexes must be unlocked in a lock-reverse order. GenericMutexGuard::drop might panic if this is violated.

Example

See StaticMutex.

Implementations

impl<System, T: 'static> GenericMutex<Hunk<System, UnsafeCell<MaybeUninit<T>>>, MutexRef<'static, System>> where
System: KernelMutex + KernelStatic,

pub const fn define() -> Definer<System, DefaultSource<T>>

Construct a Definer to define a mutex in a configuration function.

impl<Cell, Mutex, T> GenericMutex<Cell, Mutex> where
Cell: Deref<Target = UnsafeCell<MaybeUninit<T>>>,
Mutex: MutexHandle,

pub fn lock(&self) -> LockResult<GenericMutexGuard<'_, Cell, Mutex>>

Acquire the mutex, blocking the current thread until it is able to do so.

pub fn try_lock(&self) -> TryLockResult<GenericMutexGuard<'_, Cell, Mutex>>

Attempt to acquire the mutex.

pub fn mark_consistent(&self) -> Result<(), MarkConsistentError>

Mark the state protected by the mutex as consistent.

pub fn get_ptr(&self) -> *mut T

Get a raw pointer to the contained data.