Struct r3::time::Time

#[repr(transparent)]pub struct Time { /* fields omitted */ }

Represents a timestamp used by the API surface of the R3 RTOS.

The origin is application-defined. If an application desires to represent a calender time using Time, it’s recommended to use the midnight UTC on January 1, 1970 (a.k.a. “UNIX timestamp”) as the origin.

Time is backed by u64 and can represent up to 213,503,982 days with microsecond precision.

Implementations

impl Time

pub const ZERO: Self

Zero (the origin).

pub const MAX: Self

The large representable timestamp.

pub const fn from_micros(micros: u64) -> Self

Construct a new Time from the specified number of microseconds.

pub const fn from_millis(millis: u64) -> Self

Construct a new Time from the specified number of milliseconds.

Pancis if millis overflows the representable range of Time.

pub const fn from_secs(secs: u64) -> Self

Construct a new Time from the specified number of seconds.

Pancis if secs overflows the representable range of Time.

pub const fn as_micros(self) -> u64

Get the total number of whole microseconds contained in the time span between this Time and Self::ZERO.

pub const fn as_millis(self) -> u64

Get the total number of whole milliseconds contained in the time span between this Time and Self::ZERO.

pub const fn as_secs(self) -> u64

Get the total number of whole seconds contained in the time span between this Time and Self::ZERO.

pub const fn as_secs_f64(self) -> f64

Get the total number of seconds contained in the time span between this Time and Self::ZERO as f64.

Examples

use r3::time::Time;

let dur = Time::from_micros(1_201_250_000);
assert_eq!(dur.as_secs_f64(), 1201.25);

pub const fn as_secs_f32(self) -> f32

Get the total number of seconds contained in the time span between this Time and Self::ZERO as f32.

Examples

use r3::time::Time;

let dur = Time::from_micros(1_201_250_000);
assert_eq!(dur.as_secs_f32(), 1201.25);

pub const fn core_duration_since_origin(self) -> Duration

Get the duration since the origin as ::core::time::Duration.

pub const fn core_duration_since(self, reference: Self) -> Option<Duration>

Get the duration since the specified timestamp as ::core::time::Duration. Returns None if self < reference.

pub fn duration_since(self, reference: Self) -> Option<Duration>

Get the duration since the specified timestamp as Duration. Returns None if the result overflows the representable range of Duration.

pub const fn wrapping_add(&self, duration: Duration) -> Self

Advance the time by duration and return the result.

pub const fn wrapping_sub(&self, duration: Duration) -> Self

Put back the time by duration and return the result.