Trait r3::utils::binary_heap::BinaryHeap

pub trait BinaryHeap: VecLike {
    fn heap_pop(
        &mut self, 
        ctx: impl BinaryHeapCtx<Self::Element>
    ) -> Option<Self::Element>;

    fn heap_remove(
        &mut self, 
        i: usize, 
        ctx: impl BinaryHeapCtx<Self::Element>
    ) -> Option<Self::Element>;

    fn heap_push(
        &mut self, 
        item: Self::Element, 
        ctx: impl BinaryHeapCtx<Self::Element>
    ) -> usize;
}

Min-heap.

Required methods

fn heap_pop(
&mut self,
ctx: impl BinaryHeapCtx<Self::Element>
) -> Option<Self::Element>

Remove the least item from the heap and return it.

fn heap_remove(
&mut self,
i: usize,
ctx: impl BinaryHeapCtx<Self::Element>
) -> Option<Self::Element>

Remove the item at the specified position and return it.

fn heap_push(
&mut self,
item: Self::Element,
ctx: impl BinaryHeapCtx<Self::Element>
) -> usize

Push an item onto the heap and return its position.

Loading content...