Trait r3::sync::source::Source

pub trait Source<System> {
    type Target: 'static;

    // Required method
    fn into_unsafe_cell_hunk<C>(
        self,
        cfg: &mut Cfg<'_, C>
    ) -> Hunk<System, UnsafeCell<MaybeUninit<Self::Target>>>
       where C: CfgBase<System = System>;
}

Available on crate feature sync only.

Abstracts the various ways to provide an initial value to static cell types (e.g., Mutex). It’s mostly used through their definer types.

Stability

This trait is unstable.

Required Associated Types

type Target: 'static

Required Methods

fn into_unsafe_cell_hunk<C>( self, cfg: &mut Cfg<'_, C> ) -> Hunk<System, UnsafeCell<MaybeUninit<Self::Target>>>where C: CfgBase<System = System>,

Construct a Hunk to store the value.

Self::Target in the returned Hunk is safe to access from executable objects through MaybeUninit. The client may grant mutable access to Self::Target through UnsafeCell while observing the borrow rules.